refact: Refatoração uso de BigDecimal e melhorias de usabilidade
- Refatoração Geral (refact):
  - Alteração de todos os tipos de dados de Double para BigDecimal nos models (Crypto, CryptoFromApi, Transaction, Wallet). Para evitar problemas de arredondamento.

- Atividades (refact e feat):
  - CryptoDetailsActivity, CryptoBuyFormActivity, CryptoSellFormActivity, WalletFormActivity, WalletDetailsActivity: Refatoração para suportar BigDecimal, e exibição de 8 casas apos a virgula para cryptos.

- Adapters (refact):
  - CoinWalletAdapter e WalletListAdapter: Atualizados para suportar BigDecimal.

- Tratamento de Erros (refact):
  - Melhoria no tratamento de erros em todo o aplicativo, incluindo funções suspensas com retorno de Result<Unit> e tratamento de exceções com IOException para lidar com possíveis falhas de API e banco de dados.
  - Implementação de verificações e manipulação segura de dados nulos ou incorretos de chamadas de API.

+++ app/src/main/java/br/com/povengenharia/simuladorcarteiracrypto/model/Transaction.kt
@@ -2,16 +2,17 @@ package br.com.povengenharia.simuladorcarteiracrypto.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity
 data class Transaction(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val type: TransactionType,
-    val amount: Double,
+    val amount: BigDecimal,
     val walletId: Int,
     val cryptoId: String? = null,
-    val pricePerUnit: Double? = null,
-    val quantity: Double? = null,
+    val pricePerUnit: BigDecimal? = null,
+    val quantity: BigDecimal? = null,
     val timestamp: Long = System.currentTimeMillis()
 )
